moved DisplayBoard.
