+++ src/pages/Jacdac.tsx
@@ -6,7 +6,7 @@ import React, { FC, useState, useEffect } from 'react'
 import { JDBus, JDDevice, SRV_ACCELEROMETER, REPORT_UPDATE, throttle, startDevTools, inIFrame } from 'jacdac-ts'
 import { useServices, useChange, useBus } from 'react-jacdac'
 import { Button } from '@mui/material'
-import { DisplayBoard } from '../sonification/DisplayBoard'
+import { DisplayBoard } from '../sonification/displays/DisplayBoard'
 import { JacdacProvider } from 'react-jacdac'
 import { bus } from '../bus'
 import { DataSource } from '../sonification/DataSource'
