+++ src/views/demos/DemoSimple.tsx
@@ -2,7 +2,7 @@ import React from 'react'
 import { DataSource } from '../../sonification/DataSource'
 import { Datum } from '../../sonification/Datum'
 import { NoiseSonify } from '../../sonification/displays/NoiseSonify'
-import { DisplayBoard } from '../../sonification/DisplayBoard'
+import { DisplayBoard } from '../../sonification/displays/DisplayBoard'
 import { FilterRangeTemplate } from '../../sonification/templates/FilterRangeTemplate'
 import { NoteTemplate } from '../../sonification/templates/NoteTemplate'
 import { IDemoView } from './IDemoView'
